fat: attempt (unsuccessfully) to remove the unsafe { } for the nested RefCell problem by using Cells
This does not work since the borrow checker correctly complains that the function that's run can use the mutable references in the output it produces.
This is exactly what our use of unsafe here actually enables except in our case it's (probably) safe; we're returning a reference to only a RefCell'ed entity.